v1_31_0

+++ boost/date_time/c_local_time_adjustor.hpp
@@ -1,9 +1,12 @@
- 
 #ifndef DATE_TIME_C_LOCAL_TIME_ADJUSTOR_HPP__
 #define DATE_TIME_C_LOCAL_TIME_ADJUSTOR_HPP__
-/* Copyright (c) 2002 CrystalClear Software, Inc.
- * Disclaimer & Full Copyright at end of file
- * Author: Jeff Garland 
+
+/* Copyright (c) 2002,2003 CrystalClear Software, Inc.
+ * Use, modification and distribution is subject to the 
+ * Boost Software License, Version 1.0. (See accompanying
+ * file LICENSE-1.0 or http://www.boost.org/LICENSE-1.0)
+ * Author: Jeff Garland, Bart Garst
+ * $Date: 2003/12/03 01:24:22 $
  */
 
 /*! @file c_local_time_adjustor.hpp
@@ -21,8 +24,7 @@ namespace date_time {
    *  machine are correct.  This can be a very dangerous assumption.
    */
   template<class time_type>
-  class c_local_adjustor
-  {
+  class c_local_adjustor {
   public:
     typedef typename time_type::time_duration_type time_duration_type;
     typedef typename time_type::date_type date_type;
@@ -39,14 +41,14 @@ namespace date_time {
       time_duration_type td = t.time_of_day();
       std::time_t t2 = dd.days()*86400 + td.hours()*3600 + td.minutes()*60 + td.seconds();
       std::tm* tms = std::localtime(&t2);
-      date_type d(tms->tm_year + 1900,
-                  tms->tm_mon + 1, 
-                  tms->tm_mday);
+      date_type d(static_cast<unsigned short>(tms->tm_year + 1900),
+                  static_cast<unsigned short>(tms->tm_mon + 1),
+                  static_cast<unsigned short>(tms->tm_mday));
       time_duration_type td2(tms->tm_hour,
                              tms->tm_min,
-			     tms->tm_sec,
+                             tms->tm_sec,
                              t.time_of_day().fractional_seconds());
-
+      
       return time_type(d,td2);
     }
   };
@@ -56,17 +58,5 @@ namespace date_time {
 } } //namespace date_time
 
 
-/* Copyright (c) 2002
- * CrystalClear Software, Inc.
- *
- * Permission to use, copy, modify, distribute and sell this software
- * and its documentation for any purpose is hereby granted without fee,
- * provided that the above copyright notice appear in all copies and
- * that both that copyright notice and this permission notice appear
- * in supporting documentation.  CrystalClear Software makes no
- * representations about the suitability of this software for any
- * purpose.  It is provided "as is" without express or implied warranty.
- *
- */
 
 #endif
